\name{idx2r-package}
\alias{idx2r-package}
\alias{idx2r}
\docType{package}
\title{
\packageTitle{idx2r}
}
\description{
\packageDescription{idx2r}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{idx2r}
\packageIndices{idx2r}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{idx2r}

Maintainer: \packageMaintainer{idx2r}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
